LPD-59060 Inline TermsFilterTranslator/TermsFilterTranslatorImpl
