--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/filter/TermsFilterTranslatorImpl.java
@@ -1,66 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.filter;
-
-import com.liferay.petra.string.CharPool;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.search.filter.TermsFilter;
-import com.liferay.portal.kernel.util.StringUtil;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.apache.lucene.index.Term;
-import org.apache.lucene.search.BooleanClause;
-import org.apache.lucene.search.BooleanQuery;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.search.TermInSetQuery;
-import org.apache.lucene.util.BytesRef;
-import org.apache.solr.client.solrj.util.ClientUtils;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Michael C. Han
- */
-@Component(service = TermsFilterTranslator.class)
-public class TermsFilterTranslatorImpl implements TermsFilterTranslator {
-
-	@Override
-	public Query translate(TermsFilter termsFilter) {
-		String field = _escape(termsFilter.getField());
-
-		List<BytesRef> bytesRefs = new ArrayList<>();
-
-		for (String value : termsFilter.getValues()) {
-			if (value.isEmpty()) {
-				value = StringPool.DOUBLE_APOSTROPHE;
-			}
-
-			Term term = new Term(field, ClientUtils.escapeQueryChars(value));
-
-			bytesRefs.add(term.bytes());
-		}
-
-		Query query = new TermInSetQuery(field, bytesRefs);
-
-		if (bytesRefs.size() == 1) {
-			return query;
-		}
-
-		BooleanQuery.Builder builder = new BooleanQuery.Builder();
-
-		builder.add(query, BooleanClause.Occur.SHOULD);
-
-		return builder.build();
-	}
-
-	private String _escape(String value) {
-		return StringUtil.replace(
-			value, CharPool.SPACE, StringPool.BACK_SLASH + StringPool.SPACE);
-	}
-
-}
